+++ {{cookiecutter.project_slug}}/tests/test_{{cookiecutter.pkg_name}}.py
@@ -1,39 +1,37 @@
-#!/usr/bin/env python
 """Tests for `{{ cookiecutter.pkg_name }}` package."""
-
-import pytest
-{% if cookiecutter.command_line_interface|lower == 'click' -%}
-from click.testing import CliRunner
+{% if cookiecutter.command_line_interface|lower == 'y' -%}
+from typer.testing import CliRunner
 
 from {{ cookiecutter.pkg_name }} import cli
 {%- endif %}
 
 
-@pytest.fixture
-def response():
-    """Sample pytest fixture.
-
-    See more at: http://doc.pytest.org/en/latest/fixture.html
-    """
-    # import requests
-    # return requests.get('https://github.com/audreyr/cookiecutter-pypackage')
+{%- if cookiecutter.command_line_interface|lower == 'y' %}
 
 
-def test_content(response):
-    """Sample pytest test function with the pytest fixture as an argument."""
-    # from bs4 import BeautifulSoup
-    # assert 'GitHub' in BeautifulSoup(response.content).title.string
-    del response
-{%- if cookiecutter.command_line_interface|lower == 'click' %}
-
-
-def test_command_line_interface():
+class TestCLI:
     """Test the CLI."""
+
     runner = CliRunner()
-    result = runner.invoke(cli.main)
-    assert result.exit_code == 0
-    assert '{{ cookiecutter.project_slug }}' in result.output
-    help_result = runner.invoke(cli.main, ['--help'])
-    assert help_result.exit_code == 0
-    assert '--help  Show this message and exit.' in help_result.output
+
+    def test_main(self) -> None:
+        """Test main call of CLI"""
+
+        result = self.runner.invoke(cli.main)
+        assert result.exit_code == 0
+        assert "{{ cookiecutter.project_slug }}" in result.output
+
+    def test_help(self) -> None:
+        """Test --help call of CLI"""
+
+        result = self.runner.invoke(cli.main, ["--help"])
+        assert result.exit_code == 0
+
+        help_lines = list(
+            filter(lambda line: "--help" in line, result.output.split("\n"))
+        )
+        assert len(help_lines) == 1
+
+        help_text = help_lines[0]
+        assert "Show this message and exit." in help_text
 {%- endif %}